remove singular add attachment step
