+++ x-pack/platform/plugins/shared/cases/common/workflows/translations.ts
@@ -328,28 +328,6 @@ export const ADD_EVENTS_STEP_DOCUMENTATION_DETAILS = i18n.translate(
   }
 );
 
-export const ADD_ATTACHMENT_STEP_LABEL = i18n.translate(
-  'xpack.cases.workflowSteps.addAttachment.label',
-  {
-    defaultMessage: 'Cases - Add attachment to case',
-  }
-);
-
-export const ADD_ATTACHMENT_STEP_DESCRIPTION = i18n.translate(
-  'xpack.cases.workflowSteps.addAttachment.description',
-  {
-    defaultMessage: 'Adds an attachment of any registered type to a case',
-  }
-);
-
-export const ADD_ATTACHMENT_STEP_DOCUMENTATION_DETAILS = i18n.translate(
-  'xpack.cases.workflowSteps.addAttachment.documentation.details',
-  {
-    defaultMessage:
-      'This step adds a single attachment to an existing case. The `attachment` shape is determined by its `type` discriminator; the YAML editor narrows the available fields once a type is chosen. Author-friendly attachment types include `comment`, `file`, alerts, and events.',
-  }
-);
-
 export const ADD_ATTACHMENTS_STEP_LABEL = i18n.translate(
   'xpack.cases.workflowSteps.addAttachments.label',
   {
@@ -368,7 +346,7 @@ export const ADD_ATTACHMENTS_STEP_DOCUMENTATION_DETAILS = i18n.translate(
   'xpack.cases.workflowSteps.addAttachments.documentation.details',
   {
     defaultMessage:
-      'This step adds multiple attachments to an existing case in a single bulk request. Each entry follows the same per-type shape as `cases.addAttachment`.',
+      'This step adds one or more attachments to an existing case in a single bulk request. Each `attachments` entry is a per-type payload whose shape is chosen by its `type` discriminator (`comment`, alerts, events, etc.); the YAML editor narrows the available fields once a type is picked.',
   }
 );
