+++ x-pack/platform/plugins/shared/cases/public/workflows/add_attachments.tsx
@@ -5,25 +5,14 @@
  * 2.0.
  */
 
-import { buildAddAttachmentStepCommonDefinition } from '../../common/workflows/steps/add_attachment';
 import { buildAddAttachmentsStepCommonDefinition } from '../../common/workflows/steps/add_attachments';
 import { selectWorkflowAttachmentSchemas } from '../../common/workflows/steps/workflow_attachment_schemas';
 import type { UnifiedAttachmentTypeRegistry } from '../client/attachment_framework/unified_attachment_registry';
 import { createPublicCaseStepDefinition } from './shared';
 
-export const getAddAttachmentStepDefinition = (registry: UnifiedAttachmentTypeRegistry) => {
-  const members = selectWorkflowAttachmentSchemas(registry.list());
-  // No authorable type registered: skip the step rather than composing an empty union.
-  if (members.length === 0) {
-    return undefined;
-  }
-  return createPublicCaseStepDefinition({
-    ...buildAddAttachmentStepCommonDefinition(members),
-  });
-};
-
 export const getAddAttachmentsStepDefinition = (registry: UnifiedAttachmentTypeRegistry) => {
   const members = selectWorkflowAttachmentSchemas(registry.list());
+  // No authorable type registered: skip the step rather than composing an empty union.
   if (members.length === 0) {
     return undefined;
   }
